requirements.txt
Flask==2.3.2
opencv-python==4.8.1.78
tensorflow==2.14.0
Keras==2.14.0
requests==2.27.1
numpy==1.23.5
flask-opencv-streamer==1.4
h5py==3.10.0
azure-storage-blob==12.18.3
pandas==1.5.2
Pillow==10.1.0
scipy==1.11.3
uuid==1.30